+++ sprint/views.py
@@ -2,6 +2,7 @@ import datetime
 from datetime import timedelta
 
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 from django.core.mail import EmailMessage
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
@@ -32,22 +33,26 @@ def create_sprint(request, pk):
     except:
         return HttpResponseRedirect('/proyecto/')
 
+    permisos = proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',flat=True)
+    if ('create_sprint' in permisos or request.user.is_staff):
+        if request.method == 'POST':
+            lider=Usuario.objects.get(pk=proyecto.lider_proyecto.pk)
+            form = SprintForm(request.POST)
+            if form.is_valid():
+                sprint = Sprint()
+                sprint.proyecto = proyecto
+                sprint.nombre = form.cleaned_data['nombre']
+                sprint.fecha_inicio = form.cleaned_data['fecha_inicio']
+                sprint.save()
 
-    if request.method == 'POST':
-        lider=Usuario.objects.get(pk=proyecto.lider_proyecto.pk)
-        form = SprintForm(request.POST)
-        if form.is_valid():
-            sprint = Sprint()
-            sprint.proyecto = proyecto
-            sprint.nombre = form.cleaned_data['nombre']
-            sprint.fecha_inicio = form.cleaned_data['fecha_inicio']
-            sprint.save()
-
-            return HttpResponseRedirect('/proyecto/' + str(proyecto.id))
+                return HttpResponseRedirect('/proyecto/' + str(proyecto.id))
+        else:
+            form = SprintForm()
+        return render(request, 'sprint_create.html', {'form': form})
     else:
-        form = SprintForm()
+        raise PermissionDenied
 
-    return render(request, 'sprint_create.html', {'form': form})
+
 
 
 @login_required(None, 'login', '/login/')
@@ -64,33 +69,37 @@ def detail_sprint(request, pk):
     except:
         return HttpResponseRedirect('/proyecto/')
 
-    if sprint.duracion > 0:
-        sprint.fecha_fin = sprint.fecha_inicio + timedelta(days=sprint.duracion)
+    permisos = sprint.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename', flat=True)
+    if ('create_sprint' in permisos or request.user.is_staff):
+        if sprint.duracion > 0:
+            sprint.fecha_fin = sprint.fecha_inicio + timedelta(days=sprint.duracion)
 
-    # Aca verificamos si ya inicio el sprint
+        # Aca verificamos si ya inicio el sprint
 
-    if (datetime.date.today() <= sprint.fecha_inicio) and sprint.estado_sprint == 'PEN':
-        sprint.estado_sprint = 'PEN'
-        print 'Proyecto pendiente'
+        if (datetime.date.today() <= sprint.fecha_inicio) and sprint.estado_sprint == 'PEN':
+            sprint.estado_sprint = 'PEN'
+            print 'Proyecto pendiente'
+        else:
+            print 'Proyecto iniciado'
+            sprint.estado_sprint = 'INI'
+
+        # Aca verificamos si ya finalizo
+        if sprint.fecha_fin != None:
+            if sprint.fecha_fin <= datetime.date.today():
+                print 'Proyecto finalizado'
+                sprint.estado_sprint = 'FIN'
+        sprint.save()
+
+        uss = sprint.uss.all()
+        tipos_us = uss.values('tipoUS').distinct()
+        tipos = []
+        for t in tipos_us:
+            tipo_id = t["tipoUS"]
+            tipos.append([TipoUS.objects.get(id=tipo_id), uss.filter(tipoUS=tipo_id)])
+
+        return render(request, 'sprint_detail.html', {'sprint': sprint, 'tipos': tipos})
     else:
-        print 'Proyecto iniciado'
-        sprint.estado_sprint = 'INI'
-
-    # Aca verificamos si ya finalizo
-    if sprint.fecha_fin != None:
-        if sprint.fecha_fin <= datetime.date.today():
-            print 'Proyecto finalizado'
-            sprint.estado_sprint = 'FIN'
-    sprint.save()
-
-    uss = sprint.uss.all()
-    tipos_us = uss.values('tipoUS').distinct()
-    tipos = []
-    for t in tipos_us:
-        tipo_id = t["tipoUS"]
-        tipos.append([TipoUS.objects.get(id=tipo_id), uss.filter(tipoUS=tipo_id)])
-
-    return render(request, 'sprint_detail.html', {'sprint': sprint, 'tipos': tipos})
+        raise PermissionDenied
 
 
 @login_required(None, 'login', '/login/')
@@ -109,25 +118,30 @@ def asignar_us(request, pk):
     except:
         return HttpResponseRedirect('/proyecto/')
 
-    uss = US.objects.filter(Q(proyecto=sprint.proyecto, sprint=None) | Q(sprint=sprint))
+    permisos = sprint.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',
+                                                                                               flat=True)
+    if ('asignar_us' in permisos or request.user.is_staff):
+        uss = US.objects.filter(Q(proyecto=sprint.proyecto, sprint=None) | Q(sprint=sprint))
 
-    if request.method == 'POST':
-        form = AsignarUSForm(request.POST)
-        form.fields["uss"].queryset = uss
-        if form.is_valid():
-            sprint.uss = form.cleaned_data['uss']
-            acumulador = 0
-            for us in sprint.uss.all():
-                acumulador += us.tiempo_planificado
-            sprint.duracion = acumulador
-            sprint.save()
+        if request.method == 'POST':
+            form = AsignarUSForm(request.POST)
+            form.fields["uss"].queryset = uss
+            if form.is_valid():
+                sprint.uss = form.cleaned_data['uss']
+                acumulador = 0
+                for us in sprint.uss.all():
+                    acumulador += us.tiempo_planificado
+                sprint.duracion = acumulador
+                sprint.save()
 
-            return HttpResponseRedirect('/sprint/' + str(sprint.id))
+                return HttpResponseRedirect('/sprint/' + str(sprint.id))
+        else:
+            form = AsignarUSForm(initial={'uss': sprint.uss.all})
+            form.fields["uss"].queryset = uss
+
+        return render(request, 'asignar_us.html', {'form': form, 'sprint': sprint})
     else:
-        form = AsignarUSForm(initial={'uss': sprint.uss.all})
-        form.fields["uss"].queryset = uss
-
-    return render(request, 'asignar_us.html', {'form': form, 'sprint': sprint})
+        raise PermissionDenied
 
 
 @login_required(None, 'login', '/login/')
@@ -141,10 +155,13 @@ def borrar_sprint(request, pk):
         sprint = Sprint.objects.get(pk=pk)
     except:
         return HttpResponseRedirect('/proyecto/')
+    permisos = sprint.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename', flat=True)
+    if ('delete_sprint' in permisos or request.user.is_staff):
+        sprint.delete()
 
-    sprint.delete()
-
-    return HttpResponseRedirect('/proyecto/')
+        return HttpResponseRedirect('/proyecto/')
+    else:
+        raise PermissionDenied
 
 
 @login_required(None, 'login', '/login/')
@@ -166,18 +183,22 @@ def modificar_sprint(request, pk):
     except:
         return HttpResponseRedirect('/proyecto/')
 
-    if request.method == 'POST':
-        form = SprintForm(request.POST)
-        if form.is_valid():
-            sprint.nombre = form.cleaned_data['nombre']
-            sprint.fecha_inicio = form.cleaned_data['fecha_inicio']
-            sprint.save()
-            return HttpResponseRedirect('/sprint/' + str(sprint.id))
-    else:
-        form = SprintForm(initial={'nombre': sprint.nombre,
-                                   'fecha_inicio': sprint.fecha_inicio})
+    permisos = sprint.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',flat=True)
+    if ('delete_sprint' in permisos or request.user.is_staff):
+        if request.method == 'POST':
+            form = SprintForm(request.POST)
+            if form.is_valid():
+                sprint.nombre = form.cleaned_data['nombre']
+                sprint.fecha_inicio = form.cleaned_data['fecha_inicio']
+                sprint.save()
+                return HttpResponseRedirect('/sprint/' + str(sprint.id))
+        else:
+            form = SprintForm(initial={'nombre': sprint.nombre,
+                                       'fecha_inicio': sprint.fecha_inicio})
 
-        return render(request, 'sprint_create.html', {'form': form, 'sprint': sprint})
+            return render(request, 'sprint_create.html', {'form': form, 'sprint': sprint})
+    else:
+        raise PermissionDenied
 
 
 @login_required(None, 'login', '/login/')
@@ -192,12 +213,19 @@ def iniciar_sprint(request, pk):
     """
     try:
         sprint = Sprint.objects.get(pk=pk)
+    except:
+        return HttpResponseRedirect('/proyecto/')
+
+    permisos = sprint.proyecto.equipos.filter(usuarios=request.user.id).distinct().values_list('permisos__codename',
+                                                                                               flat=True)
+    if ('iniciar_sprint' in permisos or request.user.is_staff):
         sprint.fecha_inicio = datetime.date.today()
         sprint.fecha_fin = sprint.fecha_inicio + timedelta(days=sprint.duracion)
         print 'Proyecto iniciado'
         sprint.estado_sprint = 'INI'
         sprint.save()
-    except:
-        return HttpResponseRedirect('/proyecto/')
+        return HttpResponseRedirect('/sprint/' + str(sprint.id))
+    else:
+        raise PermissionDenied
 
-    return HttpResponseRedirect('/sprint/' + str(sprint.id))
+
